Creates a cluster running the application
TL;DR
-----

Implements initial application install/cluster creation

Details
-------

Builds on the upstream code by creating an EC2 instance and
installing the application onto it using the Replicated
Embedded Cluster. The instance will be configured with a
default password for the Admin Console and can be accessed
via SSH and the Admin Console from anywhere on the Internet.

Notes
-----

* Currently the default password for the admin console is
  static to your creation of the template. This means that
  your password won't be the same as everyone else's, but
  it will be the same password for all instances of your
  application.
* There is no check to determine if the Admin Console is
  ready before the stack completes. You need to check the
  Admin Console until it response. This isn't an ideal
  state and I'm exploring a few different ways to report
  readiness.