Generates unique password per install
TL;DR
-----

Moves admin console password generate from Terraform to
CloudFormation

Details
-------

Addresses the limitation of all instance having the same default
password for the Admin Console. It was limited before because the
Terraform templates generated the password and the user-data set
the value. Now the password is treated as a custom resource in
CloudFormation.

The CloudFormation custom resource uses the Python petname
library to generate the password, then saves it to AWS Secrets
Manager. This also means the password is not explicitly written
out as a stack output.

Once the isntance is running, the user can get the Admin Console
password from Secrets Manager and use it at the URL for the
console that is part of the stack outputs.