Waits for Admin Console before stack is ready (#4)
TL;DR
-----

Updates user data to signal CloudFormation when Admin Console
is ready

Details
-------

Resolves the remaining limitation of the CloudFormation install
where cluster readiness was unclear. This change makes use of the
EC2 creation policy in CloudFormation to wait on a signal from
the instance. That signal comes from the instance user data and
is sent after the cluster install completes (so Admin Console is
ready) and the password is changed (so the generated password is
the right one.