Provides a VPC for the cluster (#5)
TL;DR
-----

Places cluster in a new VPC instead of using the default

Details
-------

Creates a VPC for the cluster. Amazon reommends/requires not
installing into the default VPC, and that's what I was doing
first. This may not last, I may decided in a feature version
to prompt for a VPC to install into.