Refactors around using SNS for the license custom resource (#6)
TL;DR
-----

Creates the Replicated license using SNS vs. a direct call

Details
-------

Simplifies the architecture and splits responsibilities between customer and
vendor AWS accounts.  The original motivation for this refactoring was to keep
logic for creating a new customer/license under the vendor account.

I accomplished that first goal by switching the custom resource for the
Replicated license to use an SNS topic vs. a direct lambda call. With that
approach I can open up the topic for access from subscriber accounts and keep
the credentials to access the Replicated Vendor Portal as a secret in the
vendor's account.

Along the way, I started refactoring the second custom resource for the admin
console password and realized that AWS Secret Manager could generate the secret
for me. That let me drop the second custom resource and rely on the built-in
generation capability. This simplifies the moving parts follows the rule of
thumb of "depend on upstream".

While I was wrapping up the refactoring I made a couple of smaller tweaks as
well. One was to clean up the instance types to a smaller and more current set,
while the other was to switch up the naming for the customer and lean on the
custom ID field for the Replicated customer to link to the AWS acocunt.